+++ teclado.h
@@ -3,11 +3,6 @@
 #include "pico/stdlib.h"
 #include "hardware/gpio.h"
 #include "hardware/clocks.h"
-//#include "pio_config.h"
-
-// Biblioteca gerada pelo arquivo ws2818b.pio 
-//#include "ws2818b.pio.h"
-
 
 
 // Matrizes com nomes de colunas e linhas - GPIO - BitDogLab
